Chess-AI/requirements.txt
pygame>=2.0.0
python-chess>=0.31.4
numpy>=1.19.5